+++ test/test.js
@@ -91,9 +91,17 @@ var tests = [{
     { secondColor: '#fff', firstColor: 'white' }
   ]
 }, {
-  message: 'identical but different length hexes',
+  message: 'identical but different length hexes with default setting',
+  fixture: '.foo { color: #888; } .bar { color: #888888; }',
+  warnings: 1,
+  warningsColors: [
+    { secondColor: '#888888', firstColor: '#888' }
+  ]
+}, {
+  message: 'identical but different length hexes with `allowEquivalentNotation`',
   fixture: '.foo { color: #888; } .bar { color: #888888; }',
   warnings: 0,
+  options: { allowEquivalentNotation: true }
 }];
 
 tests.forEach(function (test) {